move more PrunedStore definitions into pruned.rs
